Bug fix in has() method
Terms were inverted on ternary operator.

The method would only work properly if the passed arguments were a string or a single element array. If an array with multiple keys were to be passed, the foreach function would only iterate through the first element of the array, thus failing to match other keys.

This could be fixed by negating the result of the ternary operator, as well.